A slightly different approach to unit tests
Instead of reconstructing the plugin object from the ground up,
the mock plugin extends the default plugin object and stubs out what
it doesn't need.

Also removed Gulp for now, since we're using the default Jasmine setup.

Also simplified the sendMessage function - instead of sending it to a
particular listener, it sends to the overall plugin.