Add some RSS tests, save some properties in the plugin namespace
In the quote and RSS plugins, the DB variable has been saved as a
property of the plugin object instead of using the 'memory' object.
This could lead to conflicts in the future, but is OK for now.
The RSS fetchFeed function has also been saved in the plugin namespace,
which allows us to spy on it easily during testing.